Add private action `privateRemoveBlocks` to hide extended interface
